floorSize almost complete, speed remains unchanged
Remember : (actual number / 5) * (floorSize/2048)

+++ Classes/Atrezzo.h
@@ -8,7 +8,7 @@ class Atrezzo : public Entity
 
 public:
 
-	int floorSize;
+	float floorSize;
 	//int _life;
 
 	Atrezzo(void);
