+++ Classes/HelloWorldScene.h
@@ -20,7 +20,7 @@ class HelloWorld : public cocos2d::Layer
 {
 
 public:
-	int floorSize;
+	float floorSize;
     // there's no 'id' in cpp, so we recommend returning the class instance pointer
     static cocos2d::Scene* createScene();
